main [fix] This code snippet introduces a new package `logging` that provides a st…
This code snippet introduces a new package `logging` that provides a structured logging mechanism for command-line applications. The `Logger` interface defines methods to log messages with different prefixes and formats. The `SheldonLogger` struct implements this interface by generating Sheldon Coo…
The `NewSheldonLogger` function creates an instance of `SheldonLogger` with a default set of prefixes. The `Info` method takes a command object, a format string, and variable arguments to construct and log the message. This logger can be used in command-line applications to provide informative